refact: moved functions from udp_comm_validation.hpp to .cpp. Refined content.
